renamed new card to card, changed definitions to values
