+++ backend/game/cards/card.py
@@ -1,7 +1,7 @@
 from abc import ABC
 
 
-class NewCard(ABC):
+class Card(ABC):
     def __init__(self, name, cost, effect, footnote=None):
         self.name = name
         self.cost = cost
