+++ backend/game/cards/discard_card.py
@@ -1,9 +1,9 @@
 from abc import abstractmethod
 
-from game.cards.new_card import NewCard
+from game.cards.card import Card
 
 
-class DiscardCard(NewCard):
+class DiscardCard(Card):
     @abstractmethod
     def immediate_effect(self, player_that_bought_the_card, other_players):
         raise NotImplementedError
