+++ backend/game/cards/discard_cards/victory_point_manipulation_cards/evacuation_orders.py
@@ -1,10 +1,10 @@
 from game.cards.discard_card import DiscardCard
-from game.cards.new_card import NewCard
+from game.cards.card import Card
 
 
 class EvacuationOrders(DiscardCard):
     def __init__(self):
-        NewCard.__init__(self, "Evacuation Orders", 7,
+        Card.__init__(self, "Evacuation Orders", 7,
                          "All other monsters lose 5[Star].")
 
     def immediate_effect(self, player_that_bought_the_card, other_players):
