+++ backend/game/cards/discard_cards/victory_point_manipulation_cards/gas_refinery.py
@@ -1,10 +1,10 @@
 from game.cards.discard_card import DiscardCard
-from game.cards.new_card import NewCard
+from game.cards.card import Card
 
 
 class GasRefinery(DiscardCard):
     def __init__(self):
-        NewCard.__init__(self, "Gas Refinery", 6,
+        Card.__init__(self, "Gas Refinery", 6,
                          "+ 2[Star] and other monsters lose 3[health]", "")
 
     def immediate_effect(self, player_that_bought_the_card, other_players):
