+++ backend/game/deck/deck_test.py
@@ -1,6 +1,6 @@
 import pytest
 
-import game.definitions.test_constants as constants
+import game.values.test_constants as constants
 from game.deck.deck import Deck
