+++ backend/game/player/player.py
@@ -1,5 +1,5 @@
-from game.definitions import constants
-from game.definitions.locations import Locations
+from game.values import constants
+from game.values.locations import Locations
 
 
 class Player:
