+++ backend/game/values/locations.py
